requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.9
python-dotenv==0.19.0
alembic==1.7.1
python-multipart==0.0.5
Jinja2==3.0.1
aiofiles==0.7.0